Refactor page rendering based on user profile and optimize account retrieval
Moved CategorySelection and GeneralProfileQuestionnaire into separate components, and adjusted page rendering based on the general profile availability. Extracted the getOrCreateAccount function into a separate library for code reusability. This function now gets the account if it exists, or creates a new one if it doesn't, reducing the redundancy in the codebase. Also added revalidation in the updateGeneralProfile action.