Refactor account creation and profile handling
Improved the handling of account creation by adding an optional parameter to the `getOrCreateAccount` function. This allows us to control whether or not a new account creation is mandatory. Additionally, the General Profile form handling has been extracted to a separate `GeneralProfileQuestionnaire` component for better modularization and resulted in cleaner code in `GeneralProfileView`.